[Refact] domain을 모듈로 분리
