+++ build-logic/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    `kotlin-dsl`
+    `kotlin-dsl-precompiled-script-plugins`
+}
+
+dependencies {
+    implementation(libs.kotlin.gradle.plugin)
+    implementation(libs.verify.detektPlugin)
+}
