+++ build-logic/src/main/kotlin/effectivekotlin.kotlin.library.gradle.kts
@@ -1,0 +1,16 @@
+import org.bmsk.effectivekotlin.configureKotlin
+import org.bmsk.effectivekotlin.configureTest
+import org.gradle.kotlin.dsl.kotlin
+
+plugins {
+    kotlin("jvm")
+}
+
+configureKotlin()
+configureTest()
+
+tasks {
+    test {
+        useJUnitPlatform()
+    }
+}
