+++ build-logic/src/main/kotlin/org/bmsk/effectivekotlin/Extenstion.kt
@@ -1,0 +1,9 @@
+package org.bmsk.effectivekotlin
+
+import org.gradle.api.artifacts.VersionCatalog
+import org.gradle.api.artifacts.VersionCatalogsExtension
+import org.gradle.api.plugins.ExtensionContainer
+import org.gradle.kotlin.dsl.getByType
+
+internal val ExtensionContainer.libs: VersionCatalog
+    get() = getByType<VersionCatalogsExtension>().named("libs")
