+++ build.gradle.kts
@@ -6,14 +6,13 @@ plugins {
 group = "camp.nextstep.edu"
 version = "1.0-SNAPSHOT"
 
-repositories {
-    mavenCentral()
-}
-
 dependencies {
-    testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
-    testImplementation("org.assertj", "assertj-core", "3.22.0")
-    testImplementation("io.kotest", "kotest-runner-junit5", "5.5.0")
+    implementation(projects.domain)
+    implementation(projects.model)
+
+    testImplementation(libs.junit.jupiter)
+    testImplementation(libs.assertj.core)
+    testImplementation(libs.kotest.runner.junit5)
 }
 
 tasks {
