+++ domain/src/main/kotlin/org/bmsk/effectivekotlin/domain/model/Cars.kt
@@ -1,0 +1,11 @@
+package org.bmsk.effectivekotlin.domain.model
+
+data class Cars(
+    private val cars: List<Car>
+) {
+    val first: Car
+        get() = cars.first()
+
+    val names: List<String>
+        get() = cars.map { it.name }
+}
