+++ domain/src/main/kotlin/org/bmsk/effectivekotlin/domain/policy/movement/MovementPolicy.kt
@@ -1,0 +1,9 @@
+package org.bmsk.effectivekotlin.domain.policy.movement
+
+import org.bmsk.effectivekotlin.model.Position
+
+sealed interface MovementPolicy {
+    fun move(position: Position): Position
+
+    interface RandomMovement : MovementPolicy
+}
