+++ domain/src/main/kotlin/org/bmsk/effectivekotlin/domain/random/RandomGenerator.kt
@@ -1,0 +1,5 @@
+package org.bmsk.effectivekotlin.domain.random
+
+interface RandomGenerator {
+    fun generate(): Int
+}
