+++ domain/src/main/kotlin/org/bmsk/effectivekotlin/domain/random/ZeroToNineRandomGenerator.kt
@@ -1,0 +1,7 @@
+package org.bmsk.effectivekotlin.domain.random
+
+internal class ZeroToNineRandomGenerator : RandomGenerator {
+    override fun generate(): Int {
+        return (0..9).random()
+    }
+}
