+++ model/src/main/kotlin/org/bmsk/effectivekotlin/model/CarShape.kt
@@ -1,0 +1,5 @@
+package org.bmsk.effectivekotlin.model
+
+enum class CarShape(val symbol: String) {
+    BAR_SHAPE("-"),
+}
