+++ model/src/main/kotlin/org/bmsk/effectivekotlin/model/Position.kt
@@ -1,0 +1,5 @@
+package org.bmsk.effectivekotlin.model
+
+data class Position(
+    val xPos: Int,
+)
