--- src/main/kotlin/org/bmsk/racingcar/domain/policy/MovementPolicy.kt
@@ -1,9 +1,0 @@
-package org.bmsk.racingcar.domain.policy
-
-import org.bmsk.racingcar.model.Position
-
-sealed interface MovementPolicy {
-    fun move(position: Position): Position
-
-    interface RandomMovement : MovementPolicy
-}
