--- src/main/kotlin/org/bmsk/racingcar/domain/random/RandomGenerator.kt
@@ -1,5 +1,0 @@
-package org.bmsk.racingcar.domain.random
-
-interface RandomGenerator {
-    fun generate(): Int
-}
