--- src/main/kotlin/org/bmsk/racingcar/domain/random/ZeroToNineRandomGenerator.kt
@@ -1,7 +1,0 @@
-package org.bmsk.racingcar.domain.random
-
-internal class ZeroToNineRandomGenerator : RandomGenerator {
-    override fun generate(): Int {
-        return (0..9).random()
-    }
-}
