--- src/main/kotlin/org/bmsk/racingcar/model/CarShape.kt
@@ -1,5 +1,0 @@
-package org.bmsk.racingcar.model
-
-enum class CarShape(val symbol: String) {
-    BAR_SHAPE("-"),
-}
