--- src/main/kotlin/org/bmsk/racingcar/model/Position.kt
@@ -1,5 +1,0 @@
-package org.bmsk.racingcar.model
-
-data class Position(
-    val xPos: Int,
-)
